+++ lib/hcloud/entities/public_net.rb
@@ -2,7 +2,7 @@
 
 module HCloud
   class PublicNet < Entity
-    attribute :firewalls, :firewall, array: true, default: -> { [] }
+    attribute :firewalls, :public_net_firewall, array: true, default: -> { [] }
     attribute :floating_ips, :floating_ip, array: true, default: -> { [] }
     attribute :ipv4, :ipv4
     attribute :ipv6, :ipv6